refactor: 컴포넌트, css  폴더 분리
