+++ src/components/Header.js
@@ -1,4 +1,4 @@
-import Component from "./core/Component";
+import Component from "../core/Component";
 
 class Header extends Component {
   constructor(props) {
