+++ src/components/index.js
@@ -1,0 +1,6 @@
+export { default as Button } from "./Button";
+export { default as Header } from "./Header";
+export { default as InputBox } from "./InputBox";
+export { default as Modal } from "./Modal";
+export { default as Restaurant } from "./Restaurant";
+export { default as RestaurantList } from "./RestaurantList";
