+++ src/styles/main.css
@@ -1,0 +1,2 @@
+@import "./style.css";
+@import "./common.css";
